- Utilizamos la dependencia "bcrypt" para proteger la contraseña del usuario.
- Se modifico el error "NotFound" para todas las entidades, permitiendo ingresar mensajes personalizados según sea necesario (sigue contando con un mensaje por defecto, el cual es el más común).

- Se creó el tipo "publicUser" dentro de "usuario.entity.ts", que define los atributos que se devuelven del usuario cuando este inicia sesión.

- Se creó el método "asPublicUser()" dentro de la clase "Usuario", el cual convierte al usuario en un "publicUser".

- Utilizamos la dependencia JSON Web Tokens para definir la sesión del usuario.

- Utilizamos la dependencia "cookie-parser" para guardar el token dentro de una cookie (y que la sesión del usuario sea un poco más segura).

- Adaptamos el método "findOne" de la entidad "Usuario" para que tenga en cuenta el token a la hora de permitir ser utilizado.

- LogIn implementado en back-end. CONECTAR CON EL FRONT Y PROBAR

+++ src/usuario/usuario.controller.ts
@@ -1,15 +1,48 @@
 import { Request,Response,NextFunction } from "express"
-import { Usuario } from "./usuario.entity.js"
+import { publicUser, Usuario } from "./usuario.entity.js"
 import { orm } from "../shared/db/orm.js"
 import { validarUsuario, validarUsuarioLogIn, validarUsuarioLogInSafe, validarUsuarioPatch, validarUsuarioSafe } from "./usuarios.schema.js"
 import { validarFindAll } from "../shared/validarFindAll.js"
-import { UsuarioNotFoundError, UsuarioUniqueConstraintViolation } from "../shared/errors/entityErrors/usuario.errors.js"
+import { ClienteAlreadyHasPedido, UsuarioBadRequestError, UsuarioIsNotAllowedError, UsuarioNotFoundError, UsuarioUnauthorizedError, UsuarioUniqueConstraintViolation } from "../shared/errors/entityErrors/usuario.errors.js"
 import { handleErrors } from "../shared/errors/errorHandler.js"
+import bcrypt from "bcrypt"
+import jwt from 'jsonwebtoken'
+import { SECRET_JWT_KEY } from "../shared/config.js"
 
 const em = orm.em
 
+function sanitizeUsuario(req: Request, res: Response, next: NextFunction) {
+  req.body.sanitizedInput = {
+    id: req.body.id,
+    mail: req.body.mail,
+    contrasenia: req.body.contrasenia,
+    nombre: req.body.nombre,
+    apellido: req.body.apellido,
+    telefono: req.body.telefono,
+    tipoUsuario: req.body.tipoUsuario
+  }
+  Object.keys(req.body.sanitizedInput).forEach((keys) => {
+    if(req.body.sanitizedInput[keys] === undefined) {
+      delete req.body.sanitizedInput[keys]
+    }
+  })
+  next()
+}
+
+function sanitizeLogIn(req: Request, res: Response, next: NextFunction) {
+  req.body.sanitizedLogIn = {
+    mail: req.body.mail,
+    contrasenia: req.body.contrasenia
+  }
+  next()
+}
+
 async function findAllByTipoUsuario(req:Request, res:Response) {
   try{
+    const token = req.cookies.access_token
+    if(!token) {
+      throw new UsuarioUnauthorizedError
+    }
     const { tipoUsuario } = req.query
     if(tipoUsuario){
       const tipoUsuario = (req.query.tipoUsuario as string).toLowerCase()
@@ -26,6 +59,10 @@ async function findAllByTipoUsuario(req:Request, res:Response) {
 
 async function findOne(req:Request,res:Response) {
   try{
+    const token = req.cookies.access_token
+    if(!token) {
+      throw new UsuarioUnauthorizedError
+    }
     const id = Number.parseInt(req.params.id)
     const usuario = await em.findOneOrFail(Usuario, {id},)
     res.status(200).json({message: 'Usuario encontrado', data: usuario})
@@ -33,30 +70,14 @@ async function findOne(req:Request,res:Response) {
     handleErrors(error, res)
   }
 }
-/* VER VIDEO SOBRE LOG IN PARA VER SI ESTOY ENCAMINADO
-async function addUsuarioOrGetByMailYContraseña(req:Request,res:Response) {
-  try{
-    const validoUsuario = validarUsuarioSafe(req.body)
-    const validoLogIn = validarUsuarioLogInSafe(req.body)
-    if(validoLogIn.success){
-      const logIn = validarUsuarioLogIn(req.body)
-      const usuario = await em.findOneOrFail(Usuario, {mail: logIn.mail, contrasenia: logIn.contrasenia})
-      res.status(200).json({message: 'Usuario encontrado con éxito', data: usuario})
-    } else if (validoUsuario.success){
-      const usuarioValido = validarUsuario(req.body)
-      const usuario = em.create(Usuario, usuarioValido)
-      await em.flush()
-      res.status(201).json({message: 'Usuario creado con éxito', data: usuario})
-    }
-  } catch (error:any){
-    handleErrors(error, res)
-  }
-}
-*/
 
 async function addUsuario(req: Request, res: Response){
   try {
-    const usuarioValido = validarUsuario(req.body)
+    const usuarioValido = validarUsuario(req.body.sanitizedInput)
+
+    usuarioValido.contrasenia = await bcrypt.hash(req.body.sanitizedInput.contrasenia, 10)
+    console.log(typeof usuarioValido.contrasenia)
+
     const usuario = em.create(Usuario, usuarioValido)
     await em.flush()
     res.status(201).json({message: 'Usuario creado con éxito', data: usuario})
@@ -68,8 +89,39 @@ async function addUsuario(req: Request, res: Response){
   }
 }
 
+//Recibo mail y contraseña del usuario, lo busco por su mail. Si lo encuentro, valido la contraseña. Si es válida, 
+async function logInUsuario(req: Request, res: Response) {
+  try { 
+    const mailYContraseniaValidos = validarUsuarioLogIn(req.body.sanitizedLogIn)
+    const usuario = await em.findOneOrFail(Usuario, {mail: mailYContraseniaValidos.mail}, {failHandler: () => {throw new UsuarioNotFoundError('El mail ingresado no se encuentra registrado')}})
+    const esUsuarioValido = await bcrypt.compare( mailYContraseniaValidos.contrasenia, usuario.contrasenia )
+    if(!esUsuarioValido) {
+      throw new UsuarioBadRequestError('La contraseña ingresada es incorrecta')
+    }
+    const usuarioPublico = usuario.asPublicUser()
+    const token = jwt.sign({id: usuarioPublico.id, mail: usuarioPublico.mail}, SECRET_JWT_KEY, {
+      expiresIn: '3h'
+    })
+    res.cookie('access_token', token, {
+      httpOnly: true,
+      sameSite: 'strict',
+      maxAge: 3000 * 60 * 60 // Pasaje de milisegundos a segundos * Pasaje de segundos a minutos * Pasaje de minutos a horas
+    }).status(200).json({message: 'Sesión iniciada con éxito', data: usuarioPublico})
+  } catch(error: any) {
+    handleErrors(error, res)
+  }
+}
+
+function logOutUsuario(req: Request, res: Response) {
+  res.clearCookie('access_token').status(200).json({message: 'Sesión cerrada con éxito'})
+}
+
 async function updateUsuario (req:Request,res:Response){
-  try{
+  try {
+    const token = req.cookies.access_token
+    if(!token) {
+      throw new UsuarioUnauthorizedError
+    }
     const id = Number.parseInt(req.params.id)
     const usuarioToUpdate = await em.findOneOrFail(Usuario, {id})
     let usuarioUpdated
@@ -90,14 +142,21 @@ async function updateUsuario (req:Request,res:Response){
 }
 
 async function remove (req:Request,res:Response) {
-    try {
+  try {
+    const token = req.cookies.access_token
+    if(!token) {
+      throw new UsuarioUnauthorizedError
+    }
     const id = Number.parseInt(req.params.id)
     const cliente = await em.findOneOrFail(Usuario, {id})
     await em.removeAndFlush(cliente)
     res.status(200).json({message: 'El cliente ha sido eliminado con éxito', data: cliente})
   } catch(error: any) {
+    if(error.name = 'UniqueConstraintViolationException') {
+      error = new ClienteAlreadyHasPedido
+    }
     handleErrors(error, res)
   }
 }
 
-export {findAllByTipoUsuario, findOne, addUsuario, updateUsuario, remove}
+export {findAllByTipoUsuario, findOne, addUsuario, sanitizeUsuario, logInUsuario, sanitizeLogIn, logOutUsuario, updateUsuario, remove}
